feat(images): Implement image transformation processing logic by job process
Add methods for handling image transformations including resize, crop,
rotate, grayscale, and tint. Integrate AWS S3 service for image
uploading and caching of transformed images.
